Preliminary version
This version has been tested with a DO button on an Android phone as
well as the SIS debug client.  The D3 output activation has been tested
with an LED on a breadboard.  The Particle relay shield and an
electronic strike plate have not been tested with this version but
should work with this firmware as-is.